pyproject.toml
[project]
name = "repo-runner"
version = "0.1.0"
description = "Clone GitHub repos, fetch metadata, and run commands."
requires-python = ">=3.9"

dependencies = [
    "requests>=2.31.0",
    "gitpython>=3.1.43",
    "aiohttp>=3.9.0",
    "PyYAML>=6.0.1",
]
